refactor(extensions): the transaction seam gets its own file, and the drift gate its right package
Two gate answers, both structural.

extruntime.go crossed the 500-line cap, and the honest split is the one
the ingress work created: Tx and the transaction counter are one fact
read from opposite ends — a unit opens a transaction, and the ingress
port needs to know one is open — so they move together into
extruntimetx.go.

The record-drift gate moves from the root fitness package into compose.
pkg/extension may import nothing under internal, so it cannot see the
capture envelope; the root package may not depend on a MODULE
(go-arch-lint), so it cannot see capture's activity shape. Compose
already depends on both, and is where the conversion under test lives.

Also: three decode targets and a best-effort body close carry craft
waivers with their reasons, and the strict-decode pre-scan says why it
answers nothing about a document it cannot read — the decoder's message
describes the document better than the scan could.
